# feat: benchmark + fix CRC performance