--- a/Cargo.toml
+++ b/Cargo.toml
@@ -36,13 +36,14 @@ tracing = "0.1"
 
 [dev-dependencies]
 assert_matches = "1.5"
+criterion = { version = "0.3", features = ["async_tokio"] }
 dotenv = "0.15"
 futures = "0.3"
 proptest = "1"
 proptest-derive = "0.3"
 rustls-pemfile = "0.2"
 rdkafka = "0.28"
-tokio = { version = "1.14", features = ["macros"] }
+tokio = { version = "1.14", features = ["macros", "rt-multi-thread"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 uuid = { version = "0.8", features = ["v4"] }
 
@@ -51,6 +52,10 @@ default = ["transport-tls"]
 fuzzing = []
 transport-tls = ["rustls", "tokio-rustls"]
 transport-socks5 = ["async-socks5"]
+
+[[bench]]
+name = "write_throughput"
+harness = false
 
 [workspace]
 members = [