--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,7 +21,7 @@ documentation = "https://docs.rs/rskafka/"
 async-socks5 = { version = "0.5", optional = true }
 async-trait = "0.1"
 bytes = "1.1"
-crc = "2"
+crc32c = "0.6"
 futures = "0.3"
 integer-encoding = "3"
 parking_lot = "0.11"