--- a/Cargo.toml
+++ b/Cargo.toml
@@ -63,5 +63,8 @@ members = [
     "fuzz",
 ]
 
+[profile.bench]
+debug = true
+
 [package.metadata.docs.rs]
 all-features = true