```text
sequential/rskafka      time:   [288.29 us 298.38 us 308.18 us]
                        change: [-35.194% -30.910% -26.110%] (p = 0.00 < 0.05)
                        Performance has improved.

parallel/rskafka        time:   [15.793 us 16.002 us 16.193 us]
                        change: [-51.655% -49.590% -48.139%] (p = 0.00 < 0.05)
                        Performance has improved.
```